backend/requirements-dev.txt
# Development dependencies for testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0

# Include base requirements
-r requirements.txt